+++ src/utils.ts
@@ -1,7 +1,5 @@
 'use strict';
 
-import * as fs from 'fs';
-import * as cp from 'child_process';
 import * as vscode from 'vscode';
 
 let _channel: vscode.OutputChannel;
@@ -40,24 +38,4 @@ export function showError(message: string, output?: string): void {
 
   vscode.window.showErrorMessage(`Gulp Tasks: ${message}`);
   outputError(output || message);
-}
-
-export function exists(file: string): Promise<boolean> {
-  return new Promise<boolean>((resolve, _reject) => {
-    fs.exists(file, value => {
-      resolve(value);
-    });
-  });
-}
-
-export function exec(command: string, options: cp.ExecOptions): Promise<{ stdout: string; stderr: string }> {
-  return new Promise<{ stdout: string; stderr: string }>((resolve, reject) => {
-    cp.exec(command, options, (error, stdout, stderr) => {
-      if (error) {
-        reject({ error, stdout, stderr });
-      }
-
-      resolve({ stdout, stderr });
-    });
-  });
 }