+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.3
 - Added better displaying of process output messages
+- Added support to terminate long running tasks (i.e. gulp.watch)
 
 ## 0.0.2
 - Changes needed for publishing (no functional changes)